fix(estimate-a-by-formula): prevent extra base pair at start or end
Previously, the formula always add one base pair to the start or end. This is generally fine,
but when the start or end is only one base pair long, it results in two base pairs. This causes
the read/rc to miss the isoform that it should be compatible with.